ダメージ予想関連の関数をリファクタリングし、非推奨の関数を削除。新しい関数を追加してダメージ予想を更新するロジックを整理。
